AwareContext now is separated from AppController
